feat(authentication): include response object in 'Token not found in response' errors
Before throwing the 'Token not found in response' error, the entire response is added as a key to the error. This allows clients to read the server response directly and display necessary information accordingly.

Partially resolves https://github.com/SpoonX/aurelia-authentication/issues/332 for API servers that return a valid response code on errors.